Several changes
Upgraded docker-compose files to v3.5 and changed MySQL to use a named volume.

Feature: Allow a room to have multiple encryptions - meaning you can log in
with one password and encrypt messages, log in with another and encrypt messages,
and so on. When you log back in with a password you'll see only those encrypted
messages and non-encrypted messages. Previously there was an error being thrown
so this is both a feature and a bug fix.

Added Dockerfile for composer if you don't want to install it.